requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
streamlit
pypdf
requests

